static/index.html: Convert lenslets offset to a corner-based value
Traditionally (even in LFDisplay, but not always as generated by auto-rectify),
the lenslets.offset value is at the center of the image, with lenslets grid
coordinates then being in the [-0.5,+0.5] range. To remove the auto-rectify
related troubles and make the code more natural, we convert the offset value
to one based on the [0,0] image corner coordinates.

Our code to move to the corner is not 100% fool-proof but it should be
Good Enough (tm) for now.