Eschew session variable currentFrameURL, instead frameURL() function is used
The point is that we may want to compose our image from multiple frames
from different channels in case of lightsheet data.

This breaks the #frameLink functionality; this doesn't seem very useful
to me (actually, not useful at all IMO) so I didn't bother to keep that
working, we can fix it later if needed.