computation-server.py: Proof-of-concept computation services provider
Currently, we provide just bounding box intensity plotting.

This is not integrated in simple-hdf-server since these serve different
purposes; most importantly, simple-hdf-server might not be used in
production at all, with all the data statically regenerated, but you
may want to use computation-server in the production as it performs
on-demand computation jobs that cannot be precomputed.

I chose flask instead of an approach analogous to simple-hdf-server
as that approach resulted in rather foul code and I've acquired some
flask experience meanwhile.